+++ gapis/resolve/pipeline.go
@@ -77,8 +77,8 @@ func pipelinesFor(ctx context.Context, o interface{}, p *path.Pipelines, r *path
 		}
 
 		cmd := append([]uint64{}, o.Commands.From...) // make a copy of o.Commands.From
-		for i := o.Commands.To[lastSubcommand]; i >= o.Commands.From[lastSubcommand]; i-- {
-			cmd[lastSubcommand] = i
+		for i := int64(o.Commands.To[lastSubcommand]); i >= int64(o.Commands.From[lastSubcommand]); i-- {
+			cmd[lastSubcommand] = uint64(i)
 			p := o.Commands.Capture.Command(cmd[0], cmd[1:]...).Pipelines()
 			if pl, err := pipelinesFor(ctx, cmds[o.Commands.From[0]], p, r); err != api.ErrPipelineNotAvailable {
 				return pl, err
